+++ Quality2/Entities/OrderQuality.cs
@@ -15,12 +15,12 @@ namespace Quality2.Entities
         public int Width { get; set; }
         public int MinThickness { get; set; }
         public int MaxThickness { get; set; }
-        public int TensileStrengthMD { get; set; }
-        public int TensileStrengthTD { get; set; }
+        public double TensileStrengthMD { get; set; }
+        public double TensileStrengthTD { get; set; }
         public int ElongationAtBreakMD { get; set; }
         public int ElongationAtBreakTD { get; set; }
         public double CoefficientOfFrictionS { get; set; }
-        public decimal CoefficientOfFrictionD { get; set; }
+        public double CoefficientOfFrictionD { get; set; }
         public int LightTransmission { get; set; }
         public int CoronaTreatment { get; set; }
         public int StandartQualityNameID { get; set; }
