+++ Quality2/Entities/StandartQualityFilm.cs
@@ -7,13 +7,13 @@ namespace Quality2.Entities
     {
         public int ID { get; set; }
         public int FilmID { get; set; }
-        public int ThicknessVariation { get; set; }
-        public int TensileStrengthMD { get; set; }
-        public int TensileStrengthTD { get; set; }
+        public double ThicknessVariation { get; set; }
+        public double TensileStrengthMD { get; set; }
+        public double TensileStrengthTD { get; set; }
         public int ElongationAtBreakMD { get; set; }
         public int ElongationAtBreakTD { get; set; }
         public double CoefficientOfFrictionS { get; set; }
-        public decimal CoefficientOfFrictionD { get; set; }
+        public double CoefficientOfFrictionD { get; set; }
         public int? LightTransmission { get; set; }
         public int CoronaTreatment { get; set; }
         public int StandartQualityNameID { get; set; }
